+++ src/main/java/racingcar/RacingGameFactory.java
@@ -2,6 +2,12 @@ package racingcar;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.domain.Car;
+import racingcar.domain.CarMovingCondition;
+import racingcar.domain.RacingCars;
+import racingcar.domain.RacingGame;
+import racingcar.domain.RandomCarMovingCondition;
+import racingcar.domain.TimeBaseRandomGenerator;
 
 public class RacingGameFactory {
