+++ src/main/java/racingcar/view/OutputView.java
@@ -2,7 +2,7 @@ package racingcar.view;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import racingcar.Car;
+import racingcar.domain.Car;
 
 public class OutputView {
