+++ src/test/java/racingcar/RacingGameTest.java
@@ -9,6 +9,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.Car;
+import racingcar.domain.RacingCars;
+import racingcar.domain.RacingGame;
 
 public class RacingGameTest {
